% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_gg_axis.R
\name{ro_gg_axis}
\alias{ro_gg_axis}
\title{Expand Axis Limits for Plots}
\usage{
ro_gg_axis(x, start_at_zero = TRUE)
}
\arguments{
\item{x}{Numeric vector of values to be plotted.}

\item{start_at_zero}{Logical; if TRUE (default), the lower limit of the axis
will start at 0.}
}
\value{
A numeric vector of length 2, giving the minimum and maximum
axis limits.
}
\description{
Calculates axis limits for a plot that are aligned with pretty breaks,
ensuring that all data points are within the range and the maximum value
is comfortably included, in accordance with RIVM styling.
Optionally, the lower limit can be set to zero.
}
\examples{
# Example with ggplot2
\dontrun{
library(ggplot2)
data <- data.frame(x = 1:11, y = c(11:17, 5:2))

ggplot(data, aes(x, y)) +
  geom_point() +
  scale_x_continuous(
    breaks = pretty,
    limits = ro_gg_axis
  ) +
  scale_y_continuous(
    breaks = pretty,
    limits = ro_gg_axis
  )
}
}
\seealso{
Other ggplot2:
\code{\link[=ro_gg_group_ticks]{ro_gg_group_ticks()}},
\code{\link[=ro_gg_plot]{ro_gg_plot()}},
\code{\link[=ro_gg_remove_axis_gap]{ro_gg_remove_axis_gap()}},
\code{\link[=ro_gg_save]{ro_gg_save()}},
\code{\link[=ro_gg_theme]{ro_gg_theme()}},
\code{\link[=ro_gg_y_title]{ro_gg_y_title()}}
}
\concept{ggplot2}
