% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_gg_remove_axis_gap.R
\name{ro_gg_remove_axis_gap}
\alias{ro_gg_remove_axis_gap}
\title{Remove gap between data and axes}
\usage{
ro_gg_remove_axis_gap()
}
\value{
expansion vector
}
\description{
Simplifies usage of the expansion function of ggplot2. Expansion vectors are used to add
some space between the data and the axes. For certain types of graphs you need to adjust the default
ggplot expansion values to comply with the RIVM style. For example, if you create a bar plot with
ro_gg_theme(), there is a gap between the bars and the x-axis. You can remove this gap with the
\code{ro_gg_remove_axis_gap()} function. This function returns the correct expansion vector that you need to use in the
scale functions (\code{scale_y_...} or \code{scale_x_...}) of your ggplot object. You also need to adjust the
expansion if you create a line graph where the y-axis starts at 0.
}
\section{Adjust x or y scale?}{

If you create a vertical (default) bar graph, you want to remove the gap between the bars
and the x-axis, thus you need to use \code{ro_gg_remove_axis_gap()} in ggplot's \code{scale_y_...}. For horizontal
bar graphs, it is the other way around. You  want to remove the gap between the bars
and the y-axis, thus you need to use \code{ro_gg_remove_axis_gap()} in ggplot's \code{scale_x_...} function. For
line graphs where the y-axis starts at 0 you need to use \code{ro_gg_remove_axis_gap()} in ggplot's
\code{scale_y_...} function.
}

\examples{
#bar graph with gap removal
my_plot <- mtcars |>
dplyr::mutate(gear = as.factor(gear))|>
dplyr::group_by(gear) |>
dplyr::summarise(hp = mean(hp)) |>
ggplot2::ggplot(ggplot2::aes(x = gear, y = hp)) +
ggplot2::geom_bar(stat = "identity") +
ro_gg_theme() +
ggplot2::scale_y_continuous(
  expand = ro_gg_remove_axis_gap()
)

#horizontal bar graph with gap removal
my_plot2 <- mtcars |>
dplyr::mutate(gear = as.factor(gear)) |>
dplyr::group_by(gear) |>
dplyr::summarise(hp = mean(hp)) |>
ggplot2::ggplot(ggplot2::aes(x = hp, y = gear)) +
ggplot2::geom_bar(stat = "identity") +
ro_gg_theme(flip = TRUE) +
ggplot2::scale_x_continuous(
  expand = ro_gg_remove_axis_gap()
)

# line graph that starts at 0 with gap removal
my_plot3 <- mtcars |>
dplyr::group_by(gear) |>
dplyr::summarise(hp = mean(hp)) |>
ggplot2::ggplot(ggplot2::aes(x = gear, y = hp)) +
ggplot2::geom_line() +
ggplot2::expand_limits(y = 0) +
ro_gg_theme() +
ggplot2::scale_y_continuous(
  expand = ro_gg_remove_axis_gap()
)

}
\seealso{
Other ggplot2:
\code{\link[=ro_gg_axis]{ro_gg_axis()}},
\code{\link[=ro_gg_group_ticks]{ro_gg_group_ticks()}},
\code{\link[=ro_gg_plot]{ro_gg_plot()}},
\code{\link[=ro_gg_save]{ro_gg_save()}},
\code{\link[=ro_gg_theme]{ro_gg_theme()}},
\code{\link[=ro_gg_y_title]{ro_gg_y_title()}}
}
\concept{ggplot2}
