% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_gg_save.R
\name{ro_gg_save}
\alias{ro_gg_save}
\title{Save ggplot objects with an offline configuration}
\usage{
ro_gg_save(
  filename,
  plot,
  fontsize = 10.75,
  dpi = 600,
  width = 12,
  height = 8,
  units = "cm",
  ...
)
}
\arguments{
\item{filename}{Filename to save object as. Should include preferred extension (e.g "name.png")}

\item{plot}{ggplot object to save.}

\item{fontsize}{Fontsize for title and text. Default = 10.75.}

\item{dpi}{Plot resolution. Default = 600.}

\item{width, height}{Plot size in units from the units argument. Defaults are 12 and 8}

\item{units}{Possibilities: "cm", "mm", "px" or "in".}

\item{...}{Other arguments passed on to the ggsave  function.}
}
\description{
This function can help you save your ggplot objects for offline reports.
It changes the font size of your title and text and sets a default set of parameters for offline publication.
Originally made for Staat van Infectieziektes.
}
\examples{
\dontrun{
my_plot <- ggplot2::ggplot(data = datasets::mtcars, mapping = ggplot2::aes(x = wt, y = mpg)) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "mtcars plot") +
  ro_gg_theme()

  ro_gg_save(my_plot, filename = "my_plot.png")
}

}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}

Other ggplot2:
\code{\link[=ro_gg_axis]{ro_gg_axis()}},
\code{\link[=ro_gg_group_ticks]{ro_gg_group_ticks()}},
\code{\link[=ro_gg_plot]{ro_gg_plot()}},
\code{\link[=ro_gg_remove_axis_gap]{ro_gg_remove_axis_gap()}},
\code{\link[=ro_gg_theme]{ro_gg_theme()}},
\code{\link[=ro_gg_y_title]{ro_gg_y_title()}}
}
\concept{ggplot2}
