% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_gg_y_title.R
\name{ro_gg_y_title}
\alias{ro_gg_y_title}
\title{Relocate y-axis title of ggplot object to RIVM style position}
\usage{
ro_gg_y_title(plot)
}
\arguments{
\item{plot}{A ggplot object with a defined y-axis title and the custom theme \verb{ro_gg_theme(version = "1.0)} applied.}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Relocates the y-axis title of an existing ggplot object (with \code{theme_ggrivm(version = "1.0")} applied) to the right
location, following RIVM design guidelines. The function calculates the y-coordinate of the highest grid line and
positions the y-axis title slightly above this value for optimal appearance. Note: Since the title may (partly) fall
outside the plot area, always use this function in combination with ggplot's \code{coord_cartesian(clip = "off")}.
Please note: This is an experimental function. Its features and behavior may change in future updates.
}
\examples{
\dontrun{
my_plot <- ggplot2::ggplot(data = datasets::mtcars, mapping = ggplot2::aes(x = wt, y = mpg)) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "mtcars plot") +
  ROvis.utils::ro_gg_theme(version = "1.0")
my_plot

#relocate y-axis title to RIVM style position
ROvis.utils::ro_gg_y_title(my_plot) +
  ggplot2::coord_cartesian(clip = "off")
  }
}
\seealso{
Other ggplot2:
\code{\link[=ro_gg_axis]{ro_gg_axis()}},
\code{\link[=ro_gg_group_ticks]{ro_gg_group_ticks()}},
\code{\link[=ro_gg_plot]{ro_gg_plot()}},
\code{\link[=ro_gg_remove_axis_gap]{ro_gg_remove_axis_gap()}},
\code{\link[=ro_gg_save]{ro_gg_save()}},
\code{\link[=ro_gg_theme]{ro_gg_theme()}}
}
\concept{ggplot2}
