% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.ggplot2-package.R
\docType{package}
\name{ROvis.ggplot2-package}
\alias{ROvis.ggplot2}
\alias{ROvis.ggplot2-package}
\title{ROvis.ggplot2: Create data visualisations in Rijksoverheid style using ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool to uniformly visualise graphs in ggplot2 using standardized Rijksoverheid (Dutch National Government) styling. Part of the ROvis umbrella package.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
