% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.ggplot2-package.R
\docType{package}
\name{ROvis.ggplot2-package}
\alias{ROvis.ggplot2}
\alias{ROvis.ggplot2-package}
\title{ROvis.ggplot2: Rijksoverheid Visualisatie - ggplot2}
\description{
ROvis.ggplot2 isn R package that provides a comprehensive suite of fucntions to visualise graphs in ggplot2 in Rijksoverheid style.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
